Added scalastyle and scoverage plugins
