--- app/uk/gov/hmrc/vatcorrespondencedetailsfrontend/controllers/HelloWorld.scala
@@ -1,20 +1,0 @@
-package uk.gov.hmrc.vatcorrespondencedetailsfrontend.controllers
-
-import javax.inject.{Inject, Singleton}
-
-import play.api.mvc._
-
-import scala.concurrent.Future
-import play.api.i18n.{I18nSupport, MessagesApi}
-import uk.gov.hmrc.play.bootstrap.controller.FrontendController
-import uk.gov.hmrc.vatcorrespondencedetailsfrontend.config.AppConfig
-import uk.gov.hmrc.vatcorrespondencedetailsfrontend.views
-
-@Singleton
-class HelloWorld @Inject()(val messagesApi: MessagesApi, implicit val appConfig: AppConfig) extends FrontendController with I18nSupport {
-
-  val helloWorld = Action.async { implicit request =>
-    Future.successful(Ok(views.html.hello_world()))
-  }
-
-}
