--- app/uk/gov/hmrc/vatcorrespondencedetailsfrontend/views/hello_world.scala.html
@@ -1,6 +1,0 @@
-@import uk.gov.hmrc.vatcorrespondencedetailsfrontend.config.AppConfig
-@()(implicit request: Request[_], messages: Messages, appConfig: AppConfig)
-
-@main_template(title = "Hello from vat-correspondence-details-frontend", bodyClasses = None) {
-    <h1>Hello from vat-correspondence-details-frontend !</h1>
-}
