+++ conf/app.routes
@@ -1,3 +1,3 @@
 # microservice specific routes
 
-GET        /hello-world             uk.gov.hmrc.vatcorrespondencedetailsfrontend.controllers.HelloWorld.helloWorld
+GET        /hello-world             controllers.HelloWorld.helloWorld
