+++ conf/application.conf
@@ -22,7 +22,7 @@ play.modules.enabled += "uk.gov.hmrc.play.bootstrap.FrontendModule"
 play.http.filters = "uk.gov.hmrc.play.bootstrap.filters.FrontendFilters"
 
 # Custom error handler
-play.http.errorHandler = "uk.gov.hmrc.vatcorrespondencedetailsfrontend.config.ErrorHandler"
+play.http.errorHandler = "config.ErrorHandler"
 
 play.filters.headers.contentSecurityPolicy= "default-src 'self' 'unsafe-inline' localhost:9000 localhost:9032 www.google-analytics.com data:"
