+++ project/plugins.sbt
@@ -9,3 +9,7 @@ addSbtPlugin("uk.gov.hmrc" % "sbt-git-versioning" % "0.10.0")
 addSbtPlugin("uk.gov.hmrc" % "sbt-distributables" % "1.1.0")
 
 addSbtPlugin("com.typesafe.play" % "sbt-plugin" % "2.5.12")
+
+addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.5.0")
+
+addSbtPlugin("org.scalastyle" %% "scalastyle-sbt-plugin" % "0.8.0")
