+++ test/controllers/HelloWorldControllerSpec.scala
@@ -1,5 +1,20 @@
+/*
+ * Copyright 2018 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
-package uk.gov.hmrc.vatcorrespondencedetailsfrontend.controllers
+package controllers
 
 import play.api.http.Status
 import play.api.i18n.{DefaultLangs, DefaultMessagesApi}
@@ -7,7 +22,7 @@ import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import play.api.{Configuration, Environment}
 import uk.gov.hmrc.play.test.{UnitSpec, WithFakeApplication}
-import uk.gov.hmrc.vatcorrespondencedetailsfrontend.config.AppConfig
+import config.AppConfig
 
 class HelloWorldControllerSpec extends UnitSpec with WithFakeApplication {
   val fakeRequest = FakeRequest("GET", "/")
